+++ MALClient/Comm/Manga/MangaUpdateQuery.cs
@@ -8,7 +8,7 @@ namespace MALClient.Comm
     internal class MangaUpdateQuery : Query
     {
         public MangaUpdateQuery(IAnimeData item)
-            : this(item.Id, item.MyEpisodes, item.MyStatus, item.MyScore, item.MyVolumes,item.StartDate,item.EndDate)
+            : this(item.Id, item.MyEpisodes, item.MyStatus, (int)item.MyScore, item.MyVolumes,item.StartDate,item.EndDate)
         {
         }
